use the latest handler pass in by the props
Since the shouldComponentUpdate only checks content and config, if you change an event handler, the component will still be using the previous old handler. A simple fix would be to just always get the latest handler from the props to call instead of using the one stored in the closure. 

Just a suggestion, but it is a bug that I had to address in production.