Change lodash dependency to >=4.0
We have a lot of projects that are consuming this
that are already loading lodash 4.x for other dependencies.
Having this 3.x dependency here was causing us to load a 
separate lodash onto the page just for this.

By bumping this to 4.x+ it will use the same lodash
that everything else on the page is already using